fixup! Prevent adding user beyond the Codenvy license
