+++ cdec/installation-manager-core/pom.xml
@@ -35,7 +35,7 @@
         </dependency>
         <dependency>
             <groupId>com.codenvy.onpremises.wsmaster</groupId>
-            <artifactId>codenvy-hosted-license-shared</artifactId>
+            <artifactId>codenvy-hosted-api-license</artifactId>
         </dependency>
         <dependency>
             <groupId>com.codenvy.onpremises.wsmaster</groupId>
