+++ wsmaster/codenvy-hosted-api-audit/pom.xml
@@ -39,10 +39,6 @@
             <artifactId>codenvy-hosted-api-workspace</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.codenvy.onpremises.wsmaster</groupId>
-            <artifactId>codenvy-hosted-license-shared</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
         </dependency>
