+++ wsmaster/codenvy-hosted-api-license/src/main/java/com/codenvy/api/license/shared/CodenvyLicense.java
@@ -12,7 +12,7 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.license.shared;
+package com.codenvy.api.license.shared;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
