+++ wsmaster/codenvy-hosted-api-license/src/main/java/com/codenvy/api/license/shared/LicenseFeature.java
@@ -12,14 +12,15 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.license.shared;
+package com.codenvy.api.license.shared;
 
-import com.codenvy.license.shared.CodenvyLicense.LicenseType;
+import com.codenvy.api.license.shared.CodenvyLicense.LicenseType;
+import com.codenvy.api.license.shared.exception.IllegalLicenseFormatException;
 
 import java.text.ParseException;
 import java.util.IllegalFormatException;
 
-import static com.codenvy.license.shared.CodenvyLicense.EXPIRATION_DATE_FORMAT;
+import static com.codenvy.api.license.shared.CodenvyLicense.EXPIRATION_DATE_FORMAT;
 import static java.lang.String.format;
 
 /**
