+++ wsmaster/codenvy-hosted-api-license/src/main/java/com/codenvy/api/license/shared/exception/InvalidLicenseException.java
@@ -12,27 +12,17 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.api.license.server.shared.dto;
-
-import com.codenvy.api.license.server.shared.model.Issue;
-import org.eclipse.che.dto.shared.DTO;
+package com.codenvy.api.license.shared.exception;
 
 /**
- * @author Dmytro Nochevnov
+ * @author Anatoliy Bazko
  */
-@DTO
-public interface IssueDto extends Issue {
-    @Override
-    Status getStatus();
+public class InvalidLicenseException extends LicenseException {
+    public InvalidLicenseException(String message) {
+        super(message);
+    }
 
-    void setStatus(Status status);
-
-    IssueDto withStatus(Status status);
-
-    @Override
-    String getMessage();
-
-    void setMessage(String message);
-
-    IssueDto withMessage(String message);
+    public InvalidLicenseException(String message, Throwable cause) {
+        super(message, cause);
+    }
 }
