+++ wsmaster/codenvy-hosted-api-license/src/test/java/com/codenvy/api/license/shared/CodenvyLicenseFactoryTest.java
@@ -12,7 +12,7 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.license.shared;
+package com.codenvy.api.license.shared;
 
 import org.mockito.testng.MockitoTestNGListener;
 import org.testng.annotations.BeforeMethod;
