+++ wsmaster/codenvy-hosted-api-license/src/test/resources/license-public-key.properties
