+++ wsmaster/pom.xml
@@ -56,7 +56,6 @@
         <module>codenvy-hosted-integration-tests</module>
         <module>codenvy-ldap</module>
         <module>codenvy-hosted-api-license</module>
-        <module>codenvy-hosted-license-shared</module>
         <module>codenvy-hosted-report-shared</module>
         <module>codenvy-hosted-report</module>
         <module>codenvy-hosted-api-admin</module>
